+++ Sources/RunnerLib/Files Import/DangerFileGenerator.swift
@@ -7,15 +7,17 @@ public final class DangerFileGenerator {
     public func generateDangerFile(fromContent content: String, fileName: String, logger: Logger) throws {
         var dangerContent = content
         let importsRegex = NSRegularExpression.filesImportRegex
+        let range = NSRange(location: 0, length: content.count)
 
-        importsRegex.enumerateMatches(in: content, options: [], range: NSRange(location: 0, length: content.count), using: { result, _, _ in
+        importsRegex.enumerateMatches(in: content, options: [], range: range, using: { result, _, _ in
             // Adjust the result to have the correct range also after dangerContent is modified
             guard let result = result?.adjustingRanges(offset:
                 dangerContent.utf16.count - content.utf16.count) else { return }
             let url = importsRegex.replacementString(for: result, in: dangerContent, offset: 0, template: "$1")
 
             guard let fileContent = try? String(contentsOfFile: url),
-                let replacementRange = Range<String.Index>(result.range, in: dangerContent) else {
+                let replacementRange = Range<String.Index>(result.range, in: dangerContent)
+            else {
                 logger.logWarning("File not found at \(url)")
                 return
             }