+++ Tests/RunnerLibTests/MockedExecutor.swift
@@ -5,13 +5,13 @@ final class MockedExecutor: ShellExecuting {
     var receivedCommands: [String] = []
     var result: (String) -> String = { _ in "" }
 
-    func execute(_ command: String, arguments: [String], environmentVariables _: [String: String], outputFile: String?) -> String {
+    func execute(_ command: String, arguments: [String], environmentVariables _: [String: String], outputFile _: String?) -> String {
         let receivedCommand = makeReceivedCommand(command, arguments: arguments)
         receivedCommands.append(receivedCommand)
         return result(receivedCommand)
     }
 
-    func spawn(_ command: String, arguments: [String], environmentVariables _: [String: String], outputFile: String?) throws -> String {
+    func spawn(_ command: String, arguments: [String], environmentVariables _: [String: String], outputFile _: String?) throws -> String {
         let receivedCommand = makeReceivedCommand(command, arguments: arguments)
         receivedCommands.append(receivedCommand)
         return result(receivedCommand)